move errors.ts and open_project_dialog.ts into lib
