+++ app/scripts/src/lib/errors.ts
